Consolidate the validation into a single suite
