--- tests/e2e/verify-install/prometheus/prometheus_suite_test.go
@@ -1,20 +1,0 @@
-// Copyright (c) 2021, Oracle and/or its affiliates.
-// Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
-
-package prometheus_test
-
-import (
-	"fmt"
-	"testing"
-
-	"github.com/onsi/ginkgo"
-	"github.com/onsi/ginkgo/config"
-	"github.com/onsi/ginkgo/reporters"
-	"github.com/onsi/gomega"
-)
-
-func TestPrometheus(t *testing.T) {
-	gomega.RegisterFailHandler(ginkgo.Fail)
-	junitReporter := reporters.NewJUnitReporter(fmt.Sprintf("prometheus-%d-test-result.xml", config.GinkgoConfig.ParallelNode))
-	ginkgo.RunSpecsWithDefaultAndCustomReporters(t, "Prometheus Suite", []ginkgo.Reporter{junitReporter})
-}
